% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{value}

\item{fun}{function}

\item{alpha}{alpha value}

\item{cx}{cx value}

\item{...}{}
}
\value{

}
\description{
Bootstrap
}
\examples{
myboot2(iter=10000,x,fun="mean",alpha=0.05,cx=1.5)
}
