% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychisim2.R
\name{mychisim2}
\alias{mychisim2}
\title{Chi-square statistic for two populations}
\usage{
mychisim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 3,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  ymax = 0.07,
  x = 40,
  y = 0.04,
  ...
)
}
\arguments{
\item{n1}{Sample Size 1}

\item{n2}{Sample Size 2}

\item{sigma1}{Standard Deviation of Population 1}

\item{sigma2}{Standard Deviation of Population 2}

\item{mean1}{mean of poupulation 1}

\item{mean2}{mean of poupulation 2}

\item{iter}{Number of iterations}

\item{ymax}{parameter to adjust the height of the plot}

\item{x}{x coordinate of legend}

\item{y}{y coordinate of legend}

\item{...}{}
}
\value{

}
\description{
Chi-square statistic for two populations
}
\examples{
mychisim2(n1=10,n2=14,sigma1=3,sigma2=3,mean1=5,mean2=10,iter=1000,ymax=0.07,x=40,y=0.04)
}
