% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Mean confidence interval}
\usage{
myci(x = 0, conf.int = 0.95)
}
\arguments{
\item{x}{Sample}

\item{conf.int}{interval of confidence}
}
\value{

}
\description{
Mean confidence interval
}
\examples{
myci(x=rnorm(7,mean=3,sd=1),conf.int=0.95)
}
