% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{CLT Binomial}
\usage{
mycltb(n = 5, iter = 10000, p = 0.5, ...)
}
\arguments{
\item{n}{size of samples}

\item{iter}{number of iterations or samples taken}

\item{p}{probability of success}

\item{...}{parameters to modify the histogram}
}
\value{

}
\description{
CLT Binomial
}
\examples{
mycltb(n=5,iter=10000,p=0.5)
}
