% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{CLT Poisson}
\usage{
mycltp(n = 10, iter = 10000, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations or number of samples taken}

\item{...}{parameters to change histogram}
}
\value{

}
\description{
CLT Poisson
}
