% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{CLT Uniform}
\usage{
mycltu(n = 20, iter = 10000, a = 0, b = 10)
}
\arguments{
\item{n}{size of sample}

\item{iter}{number of iterations or spales taken}

\item{a}{lower limit of possible outcome}

\item{b}{upper limit of possible outcone}
}
\value{

}
\description{
CLT Uniform
}
\examples{
mycltu(n=20,iter=10000,a=0,b=10)
}
