% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Square Function}
\usage{
myf(x, y)
}
\arguments{
\item{x}{}
}
\value{
a vector of square components in the form of a vector object
}
\description{
takes vector and returns a vector of squared components
}
\examples{
x = 1:30 ; myf(x)
}
