% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFsim2.R
\name{myFsim2}
\alias{myFsim2}
\title{F simulation}
\usage{
myFsim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 2,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  ymax = 0.9,
  x = 6,
  y = 0.5,
  ...
)
}
\arguments{
\item{n1}{Sample Size 1}

\item{n2}{Sample Size 2}

\item{sigma1}{Standard Deviation of Population 1}

\item{sigma2}{Standard Deviation of Population 2}

\item{mean1}{mean of population 1}

\item{mean2}{mean of population 2}

\item{iter}{number of iterations}

\item{ymax}{parameter to adjust the height of the plot}

\item{x}{x coordinate of legend}

\item{y}{y coordinate of legend}

\item{...}{}
}
\value{

}
\description{
F simulation
}
\examples{
myFsim2(n1=10,n2=14,sigma1=3,sigma2=2,mean1=5,mean2=10,iter=1000,ymax=0.9,x=6,y=0.5)
}
