% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric Simulation}
\usage{
myhyper(iter, N, r, n)
}
\arguments{
\item{iter}{Number of Iterations}

\item{N}{Total number of elements}

\item{r}{Number of Success in the N elements}

\item{n}{Number of elements drawn}
}
\value{

}
\description{
Hypergeometric Simulation
}
\examples{
myhyper(100,10,5,8)
}
