% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximun Likelyhood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function to be applied}

\item{x}{first parameter}

\item{param}{second parameter}

\item{...}{}
}
\value{

}
\description{
Maximun Likelyhood
}
