% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{Binomial Probability}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{Number of trials until the rth success is observed}

\item{r}{Number of Successes}

\item{p}{Probability of success on a single Bernoulli trial}
}
\value{

}
\description{
Binomial Probability
}
