% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title:myncurve:}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu:}{mean}

\item{sigma:}{standard deviation}

\item{a:}{upper limit (tail)}
}
\value{

}
\description{
Title:myncurve:
}
