% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTsim.R
\name{myTsim}
\alias{myTsim}
\title{T Simulation}
\usage{
myTsim(
  n1 = 10,
  sigma1 = 3,
  mean1 = 5,
  iter = 1000,
  ymax = 0.1,
  x = 2,
  y = 0.3,
  ...
)
}
\arguments{
\item{n1}{Sample Size}

\item{sigma1}{standard deviation of population}

\item{mean1}{mean of population}

\item{iter}{number of iterations}

\item{ymax}{parameter to adjust the height of the plot}

\item{x}{x coordinate of legend}

\item{y}{y coordinate of legend}

\item{...}{}
}
\value{

}
\description{
T Simulation
}
\examples{
myTsim(n1=10,sigma1=3,mean1=5,iter=1000,ymax=0.1,x=2,y=0.3)
}
