% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTsim2.R
\name{myTsim2}
\alias{myTsim2}
\title{T simulation with 2 populations}
\usage{
myTsim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 3,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  ymax = 0.5,
  x = 2,
  y = 0.4,
  ...
)
}
\arguments{
\item{n1}{Sample Size 1}

\item{n2}{Smaple Size 2}

\item{sigma1}{Standard Deviation of Population 1}

\item{sigma2}{Standard Deviation of Population 2}

\item{mean1}{mean of population 1}

\item{mean2}{mean of population 1}

\item{iter}{number of iterations}

\item{ymax}{parameter to adjust the height of the plot}

\item{x}{x coordinate of legend}

\item{y}{y coordinate of legend}

\item{...}{}
}
\value{

}
\description{
T simulation with 2 populations
}
\examples{
myTsim2(n1=10,n2=14,sigma1=3,sigma2=3,mean1=5,mean2=10,iter=1000,ymax=0.5,x=2,y=0.4)
}
