% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquareValue.R
\name{SquareValue}
\alias{SquareValue}
\title{Square Function}
\usage{
SquareValue(x)
}
\arguments{
\item{x}{}
}
\value{
a vector of square components in the form of a vector object
}
\description{
takes vector and returns a vector of squared components
}
\examples{
x = 1:30 ; SquareValue(x)
}
