% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{Number of Interations}

\item{n}{number of trials}

\item{p}{Probability of success on a single Bernoulli trial}
}
\value{

}
\description{
Title
}
